# Accelerate adding large numbers of points