--- a/napari/benchmarks/benchmark_points_layer.py
+++ b/napari/benchmarks/benchmark_points_layer.py
@@ -37,6 +37,9 @@ class Points2DSuite:
         """Time to get current value."""
         self.layer.get_value((0,) * 2)
 
+    def time_add(self, n):
+        self.layer.add(self.data)
+
     def mem_layer(self, n):
         """Memory used by layer."""
         return self.layer