--- a/napari/layers/points/points.py
+++ b/napari/layers/points/points.py
@@ -1207,8 +1207,12 @@ class Points(Layer):
             with self.block_update_properties():
                 self.current_face_color = face_color
 
-        size_index = np.ix_(index, self._dims_displayed)
-        size = np.unique(np.mean(self.size[size_index], axis=0))
+        # Calculate the mean size across the displayed dimensions for
+        # each point to be consistent with `_view_size`.
+        mean_size = np.mean(
+            self.size[np.ix_(index, self._dims_displayed)], axis=1
+        )
+        size = np.unique(mean_size)
         if len(size) == 1:
             size = size[0]
             with self.block_update_properties():