--- a/napari/layers/utils/layer_utils.py
+++ b/napari/layers/utils/layer_utils.py
@@ -807,7 +807,9 @@ class _FeatureTable:
         if size < current_size:
             self.remove(range(size, current_size))
         elif size > current_size:
-            to_append = pd.concat([self._defaults] * (size - current_size))
+            to_append = self._defaults.reindex(
+                index=range(size - current_size), method='pad'
+            )
             self.append(to_append)
 
     def append(self, to_append: pd.DataFrame) -> None:
@@ -828,11 +830,6 @@ class _FeatureTable:
         indices : Any
             The indices of the rows to remove. Must be usable as the labels parameter
             to pandas.DataFrame.drop.
-
-        Returns
-        -------
-        pd.DataFrame
-            The resulting features table, which contain copies of the existing data.
         """
         self._values = self._values.drop(labels=indices, axis=0).reset_index(
             drop=True