# Description

This speeds up two particularly slow lines that get executed as a result of `Points.add`. In both cases, we were originally doing python-based iteration, but both cases can be replaced by more efficient numpy/pandas-specific functions that should behave identically.

## Type of change

- [x] Bug-fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] Breaking change (fix or feature that would cause existing functionality to not work as expected)
- [ ] This change requires a documentation update

# References

[Original Zulip discussion]((https://napari.zulipchat.com/#narrow/stream/212875-general/topic/Adding.20of.20points.20slow).

# How has this been tested?

- [ ] modified a test to check that dtypes are as expected for added rows (an alternative optimization broke this test)
- [ ] all existing tests pass

I also tested this manually with the usage described in the original zulip thread, where it works as expected and much faster after this PR.

I also ran the benchmark added here before and after this change. Before, we have the following results.

```
(napari-dev) ➜  napari git:(main) ✗ asv run --bench Points2DSuite.time_add --python=same
· Discovering benchmarks
· Running 1 total benchmarks (1 commits * 1 environments * 1 benchmarks)
[  0.00%] ·· Benchmarking existing-py_Users_asweet_software_miniconda3_envs_napari-dev_bin_python
[ 50.00%] ··· Running (benchmark_points_layer.Points2DSuite.time_add--).
[100.00%] ··· benchmark_points_layer.Points2DSuite.time_add                                                                                                                                                                                                           ok
[100.00%] ··· ======== =============
               param1               
              -------- -------------
                 16     3.74±0.09ms 
                 64       5.38±1ms  
                256      9.45±0.3ms 
                1024     26.5±0.5ms 
                4096     92.9±0.4ms 
               16384      376±9ms   
               65536     1.52±0.01s 
              ======== =============
```

And after we have speed ups of ~5x for large numbers of points without making things worse for small numbers.

```
(napari-dev) ➜  napari git:(add-points-faster) ✗ asv run --bench Points2DSuite.time_add --python=same
· Discovering benchmarks
· Running 1 total benchmarks (1 commits * 1 environments * 1 benchmarks)
[  0.00%] ·· Benchmarking existing-py_Users_asweet_software_miniconda3_envs_napari-dev_bin_python
[ 50.00%] ··· Running (benchmark_points_layer.Points2DSuite.time_add--).
[100.00%] ··· benchmark_points_layer.Points2DSuite.time_add                                                                                                                                                                                                           ok
[100.00%] ··· ======== ============
               param1              
              -------- ------------
                 16     3.49±0.1ms 
                 64     3.64±0.2ms 
                256     4.43±0.1ms 
                1024    7.19±0.2ms 
                4096    16.4±0.3ms 
               16384    61.7±0.4ms 
               65536     267±3ms   
              ======== ============
```

## Final checklist:
- [ ] My PR is the minimum possible work for the desired functionality
- [ ] I have commented my code, particularly in hard-to-understand areas
- [ ] I have made corresponding changes to the documentation
- [ ] I have added tests that prove my fix is effective or that my feature works
- [ ] If I included new strings, I have used `trans.` to make them localizable.
      For more information see our [translations guide](https://napari.org/docs/dev/guides/translations.html).